REF: move LazyDecoder to base.models
